Moved all analytic related functions into one file
